requirements.txt
blinker==1.7.0
certifi==2020.6.20
chardet==4.0.0
click==8.1.7
click-plugins==1.1.1
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
Flask==3.0.2
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.3
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
MarkupSafe==2.1.5
more-itertools==8.10.0
netifaces==0.11.0
oauthlib==3.2.0
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
pyserial==3.5
PyYAML==5.4.1
requests==2.25.1
six==1.16.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
wadllib==1.3.6
Werkzeug==3.0.2
XlsxWriter==3.0.2
zipp==1.0.0
